+++ src/commonnexus/tokenizer.py
@@ -239,7 +239,7 @@ def iter_words_and_punctuation(tokens, allow_punctuation_in_word=None, nexus=Non
         two punctuation characters, and the word "words".
     """
     allow_punctuation_in_word = allow_punctuation_in_word or ''
-    if not allow_punctuation_in_word and nexus is not None:
+    if nexus is not None:
         if nexus.cfg.hyphenminus_is_text:
             allow_punctuation_in_word += '-'
         if nexus.cfg.asterisk_is_text: