+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [Unreleased]
 
+- Fixed bug where the Nexus.cfg was not consulted when reading CHARACTERS MATRIX.
+
 
 ## [v1.0.0] - 2023-03-08